+++ src/installers/pipenv.ts
@@ -10,12 +10,14 @@ export default class PipenvInstaller extends BaseInstaller {
         const packageName = this.buildPackageName(name, extras);
         const versionConstraint = version ? this.buildVersionConstraint(version) : '';
         const command = version ? 'install' : 'update';
-        const cmd = `${this.pythonPath} -m pipenv ${command} ${dev ? '--dev' : ''} "${packageName}${versionConstraint}"`;
+        const installerCmd = await this.getInstallerCmd();
+        const cmd = `${installerCmd} ${command} ${dev ? '--dev' : ''} "${packageName}${versionConstraint}"`;
         await this.exec(cmd);
     }
 
     async removePackage(name: string, dev: boolean): Promise<void> {
-        const cmd = `${this.pythonPath} -m pipenv uninstall -n ${dev ? '--dev' : ''} "${name}"`;
+        const installerCmd = await this.getInstallerCmd();
+        const cmd = `${installerCmd} uninstall -n ${dev ? '--dev' : ''} "${name}"`;
         await this.exec(cmd);
     }