+++ src/sourceScanner.ts
@@ -1,10 +1,12 @@
 import { workspace, Uri } from "vscode";
+import * as fs from "fs";
 
 
 const SERVICE_RE = /(boto3|session)\.(client|resource)\(\s*['"]+(\S+)['"]+\s*\)/g;
 
 export default class SourceScanner {
     async findPythonFiles(): Promise<Uri[]> {
+        let result: Uri[] = [];
         const exclude = [
             ...Object.keys(await workspace.getConfiguration('search', null).get('exclude') || {}),
             ...Object.keys(await workspace.getConfiguration('files', null).get('exclude') || {}),
@@ -12,10 +14,23 @@ export default class SourceScanner {
             "**/typings/**",
             "**/tests/**",
         ].join(',');
-        return workspace.findFiles('**/*.py', `{${exclude}}`);
+        return await workspace.findFiles('**/*.py', `{${exclude}}`);
     }
 
-    findServices(text: string): Set<string> {
+    async readFile(file: Uri): Promise<string> {
+        return new Promise((resolve, reject) => {
+            fs.readFile(file.fsPath, { encoding: 'utf-8' }, (err, data) => {
+                if (err) {
+                    reject(err);
+                    return;
+                }
+                resolve(data);
+            });
+        });
+    }
+
+    async findServices(file: Uri): Promise<Set<string>> {
+        const text = await this.readFile(file);
         const result: Set<string> = new Set();
         let match;
         SERVICE_RE.lastIndex = 0;